+++ vignettes/examples/example.sql
@@ -1,7 +1,7 @@
 -- qname = how_many_irises
 SELECT count(*) as N FROM IRIS;
 
--- qname = short-petal-setosa
+-- qname = short_petal_setosa
 select Species, `Petal.Length`
 FROM IRIS
 WHERE Species = "setosa"