+++ NEWS.md
@@ -5,9 +5,12 @@
 * configs now inherit characteristics from earlier configs with the same name (i.e. site-wide configs may be tweaked by users) (#9)
 * renamed `runqueries()` and `runfiles()` as `run_queries()` and `run_files()`
 * added `runqueries()` and `runfiles()` as aliases for `run_queries()` and `run_files()`  
-* numerous documentation updates
+* numerous documentation updates, including:
+  - `vignette("sqlhelper")` updated
+  - `vignette("connections")` updated
 * `read_sql()` now accepts a flag to turn off cascade behaviour (#7)
 * `run_files()` no longer requires an explicit `cascade` argument 
+* `run_files()` and `run_queries()` now attempt to connect lazily using the config search path (#8)
 
 ## Bug fixes