+++ NEWS.md
@@ -2,18 +2,21 @@
 
 ## Enhancements
 
-* configs now inherit characteristics from earlier configs with the same name
+* connection configs now inherit characteristics from earlier configs with the same name
 (i.e. site-wide configs may be tweaked by users) (#9)
+* `config_examples()` added to provide example connection yaml strings
 * renamed `runqueries()` and `runfiles()` as `run_queries()` and `run_files()`
 * added `runqueries()` and `runfiles()` as aliases for `run_queries()` and
 `run_files()`
-* numerous documentation updates, including:
+* many documentation updates, including:
   - `vignette("sqlhelper")` updated
   - `vignette("connections")` updated
   - `vignette("execution")` added
+  - `vignette("use_case")` added (#14)
 * `read_sql()` now accepts a flag to turn off cascade behaviour (#7)
 * `run_files()` no longer requires an explicit `cascade` argument 
-* `run_files()` and `run_queries()` now attempt to connect lazily using the config search path (#8)
+* `run_files()` and `run_queries()` now attempt to connect lazily using the
+config search path (#8)
 
 ## Bug fixes