+++ NEWS.md
@@ -1,4 +1,4 @@
-# sqlhelper (development version)
+# sqlhelper 0.2.0
 
 ## Enhancements